fix: did not use scrubbed query params
