+++ pages/index.tsx
@@ -1,10 +1,13 @@
+import { ParsedUrlQuery } from "querystring";
 import { useRouter } from "next/router";
-import React, { useEffect } from "react";
+import React, { useEffect, useState } from "react";
 import { connect } from "react-redux";
+import { Wogaa } from "../src/components/Analytics/wogaa";
 import { Wrapper, Main } from "../src/components/Layout/Body";
 import { FooterBar } from "../src/components/Layout/FooterBar";
 import { NavigationBar } from "../src/components/Layout/NavigationBar";
 import { MainPageContainer } from "../src/components/MainPageContainer";
+
 import {
   resetCertificateState,
   retrieveCertificateByActionFailure,
@@ -17,12 +20,15 @@ interface HomePageProps {
   retrieveCertificateByActionFailure: (message: string) => void;
 }
 const HomePage: React.FunctionComponent<HomePageProps> = (props) => {
-  const router = useRouter();
+  const urlParams = useUrlParamsThenScrubUrl({ enabled: true });
+
   useEffect(() => {
+    if (urlParams === undefined) return;
+
     props.resetCertificateState();
-    if (router.query.q) {
-      const action = JSON.parse(window.decodeURI(router.query.q as string));
-      const anchorStr = decodeURIComponent(window.location.hash.substr(1));
+    if (urlParams.query.q) {
+      const action = JSON.parse(window.decodeURI(urlParams.query.q as string));
+      const anchorStr = decodeURIComponent(urlParams.uriFragment);
       const anchor = anchorStr ? JSON.parse(anchorStr) : {};
       if (action.type === "DOCUMENT") {
         props.retrieveCertificateByAction(action.payload, anchor);
@@ -30,10 +36,11 @@ const HomePage: React.FunctionComponent<HomePageProps> = (props) => {
         props.retrieveCertificateByActionFailure(`The type ${action.type} provided from the action is not supported`);
       }
     }
-  }, [props, router]);
+  }, [props, urlParams]);
 
   return (
-    <Wrapper isScrubUrlBeforeLoadingWogaa>
+    <Wrapper isLoadWogaa={false}>
+      {urlParams ? <Wogaa /> : null}
       <NavigationBar />
       <Main>
         <MainPageContainer />
@@ -48,3 +55,49 @@ export default connect(null, {
   retrieveCertificateByAction,
   retrieveCertificateByActionFailure,
 })(HomePage);
+
+type UseUrlParamsThenScrubUrlParam = {
+  /** extraction and scrubbing will only happen if this is true */
+  enabled: boolean;
+};
+type UrlParams = {
+  uriFragment: string;
+  query: ParsedUrlQuery;
+};
+const useUrlParamsThenScrubUrl = ({ enabled }: UseUrlParamsThenScrubUrlParam): UrlParams | undefined => {
+  // urlParams will only be undefined before the second render
+  // being undefined means the urlParams has not been extracted and the url has not been scrubbed yet
+  const [urlParams, setUrlParams] = useState<UrlParams | undefined>(
+    !enabled
+      ? {
+          uriFragment: "",
+          query: {},
+        }
+      : undefined
+  );
+  const router = useRouter();
+
+  useEffect(() => {
+    if (enabled) {
+      setUrlParams((currUrlParams) => {
+        // once the fragment is set, we never update it again
+        if (currUrlParams !== undefined) {
+          return currUrlParams;
+        }
+
+        const savedFragment = window.location.hash.substring(1);
+        const savedQueryParam = { ...router.query };
+
+        // scrubbbb itttttt
+        window.location.hash = "";
+        router.replace({}, undefined, { shallow: true });
+        return {
+          uriFragment: savedFragment,
+          query: savedQueryParam,
+        };
+      });
+    }
+  }, [urlParams, router, enabled]);
+
+  return urlParams;
+};
